Move signDelegate code to signMPC
