+++ packages/near-fast-auth-signer/src/utils/multi-chain/signature/types.ts
@@ -1,5 +1,4 @@
 export type RSVSignature = {
   r: string;
   s: string;
-  v: number;
 };
